Greetings, Kindly do read paragraph below!
One file is a text file (.txt) which contains the code for the task. 
Other file is a html file (.htm) which contains the visual representation of the given task and it file can be ran in any browser.
Bootstrap, CSS and JQuery are taken from cdn links hence it will only be visible in presence of an active internet connection.
Form validation is also included in the task
Username = hey@epam.com
Password = epam
Thanks and Good luck!